Move helper functions to a utils file
